content/html/content/test/test_bug583288-3.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=583288
-->
<head>
  <title>Test for Bug 583288</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload='runTests();'>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=583288">Mozilla Bug 583288</a>
<p id="display"></p>
<style>
  iframe { width: 70px; height: 40px;}
</style>
<!--<div id="content" style="display: none">-->
<div id="content">
  <iframe name='get' id='get'></iframe>
  <iframe name='post' id='post'></iframe>
  <iframe name='put' id='put'></iframe>
  <iframe name='del' id='del'></iframe>

  <form action="583288_redirect_server.sjs" target='get' method='get'>
  </form>
  <form action="583288_redirect_server.sjs" target='post' method='post'>
  </form>
  <form action="583288_redirect_server.sjs" target='put' method='put'>
  </form>
  <form action="583288_redirect_server.sjs" target='del' method='delete'>
  </form>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 583288 **/

SimpleTest.waitForExplicitFinish();

var gTestResults = {
  get:   "GET",
  post:  "POST",
  put:   "",
  del:   "",
};

var gPendingLoad = 0; // Has to be set after depending on the frames number.

function runTests()
{
  // We add a load event for the frames which will be called when the forms
  // will be submitted.
  var frames = [ document.getElementById('get'),
                 document.getElementById('post'),
                 document.getElementById('put'),
                 document.getElementById('del'),
               ];
  gPendingLoad = frames.length;

  for (var i=0; i<gPendingLoad; ++i) {
    frames[i].setAttribute('onload', "frameLoaded(this);");
  }

  // The four forms can be submitted with .submit().
  for (var i=0; i<4; ++i) {
    document.forms[i].submit();
  }
}

function frameLoaded(aFrame) {
  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
  is(aFrame.contentDocument.documentElement.textContent, gTestResults[aFrame.name],
     "cross-origin submission with redirection should work");

  if (--gPendingLoad == 0) {
    SimpleTest.finish();
  }
}

</script>
</pre>
</body>
</html>
